Remove green_box_utils, reorganize and rename things
