+++ Dockerfile
@@ -6,18 +6,18 @@ RUN apt-get -y install --no-install-recommends \
   python-pip \
   python-setuptools \
   vim \
-  nmap
+  nmap \
+  git
+
+RUN pip install wheel
 
 RUN mkdir /secondary-analysis
 WORKDIR /secondary-analysis
 
-# pip doesn't preserve the order that we ask to install things when we give a big list and
-# we need wheel before we can install anything else
-RUN pip install wheel && \
-    pip install connexion
+COPY requirements.txt .
+
+RUN pip install -r requirements.txt
 
 COPY . .
 
-RUN cd green_box_utils && pip install .
-
-CMD ["bash", "listener-start.sh"]
+CMD ["python", "lira/lira.py"]
