--- app.yaml
@@ -1,7 +1,0 @@
-service: green
-runtime: custom
-env: flex
-health_check:
-  enable_health_check: False
-manual_scaling:
-  instances: 1
