--- green-box-api.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-from green_box import app
-
-app.run(port=8080)
