--- green_box_utils/README.rst
@@ -1,13 +1,0 @@
-Overview
-========
-
-This package provides utilities used in the secondary analysis service that may be reusable elsewhere,
-with elaborate docstrings.
-
-This package provides the following:
-
-- Reusable code that actually calls the Cromwell API to start a workflow.
-
-- Reusable code that interacting with Google Cloud Storage API.
-
-- Utility functions that support listener of Green-Box.
