+++ lira/test/test_config.py
@@ -2,7 +2,9 @@
 import unittest
 import json
 from copy import deepcopy
-from green_box_utils import listener_utils
+from lira import config
+import os
+import sys
 
 
 class TestStartupVerification(unittest.TestCase):
@@ -10,64 +12,67 @@ class TestStartupVerification(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         """load the config file"""
+        # Change to test directory, as tests may have been invoked from another dir
+        dir = os.path.abspath(os.path.dirname(__file__))
+        os.chdir(dir)
         with open('config.json', 'rb') as f:
             cls.correct_test_config = json.load(f)
 
     def test_correct_config_throws_no_errors(self):
-        config = deepcopy(self.correct_test_config)
+        test_config = deepcopy(self.correct_test_config)
 
         try:  # make sure ListenerConfig executes
-            result = listener_utils.ListenerConfig(config)
+            result = config.ListenerConfig(test_config)
         except BaseException as exception:
             self.fail(
                 'ListenerConfig constructor raised an exception: {exception}'
                 .format(exception=exception))
 
         # check correct type
-        self.assertIsInstance(result, listener_utils.ListenerConfig)
+        self.assertIsInstance(result, config.ListenerConfig)
 
     def test_config_missing_required_field_throws_value_error(self):
 
         def delete_wdl_field(field_name):
             """return a config object whose first wdl config is missing a field"""
-            config = deepcopy(self.correct_test_config)
-            del config['wdls'][0][field_name]
-            return config
+            test_config = deepcopy(self.correct_test_config)
+            del test_config['wdls'][0][field_name]
+            return test_config
 
         mangled_config = delete_wdl_field('subscription_id')
-        self.assertRaises(ValueError, listener_utils.ListenerConfig, mangled_config)
+        self.assertRaises(ValueError, config.ListenerConfig, mangled_config)
         mangled_config = delete_wdl_field('workflow_name')
-        self.assertRaises(ValueError, listener_utils.ListenerConfig, mangled_config)
+        self.assertRaises(ValueError, config.ListenerConfig, mangled_config)
         mangled_config = delete_wdl_field('wdl_link')
-        self.assertRaises(ValueError, listener_utils.ListenerConfig, mangled_config)
+        self.assertRaises(ValueError, config.ListenerConfig, mangled_config)
         mangled_config = delete_wdl_field('wdl_default_inputs_link')
-        self.assertRaises(ValueError, listener_utils.ListenerConfig, mangled_config)
+        self.assertRaises(ValueError, config.ListenerConfig, mangled_config)
         mangled_config = delete_wdl_field('wdl_deps_link')
-        self.assertRaises(ValueError, listener_utils.ListenerConfig, mangled_config)
+        self.assertRaises(ValueError, config.ListenerConfig, mangled_config)
 
     def test_config_duplicate_wdl_raises_value_error(self):
 
         def add_duplicate_wdl_definition():
             """add the first wdl definition to the end of the 'wdls' json section"""
-            config = deepcopy(self.correct_test_config)
-            config['wdls'].append(config['wdls'][0])
-            return config
+            test_config = deepcopy(self.correct_test_config)
+            test_config['wdls'].append(test_config['wdls'][0])
+            return test_config
 
         mangled_config = add_duplicate_wdl_definition()
-        self.assertRaises(ValueError, listener_utils.ListenerConfig, mangled_config)
+        self.assertRaises(ValueError, config.ListenerConfig, mangled_config)
 
     def test_listener_config_exposes_all_methods_requested_in_notifications(self):
         # test that the calls made in notifications refer to attributes that
         # ListenerConfig exposes
-        config = listener_utils.ListenerConfig(self.correct_test_config)
+        test_config = config.ListenerConfig(self.correct_test_config)
         requested_listener_attributes = [
             'notification_token', 'wdls', 'cromwell_url', 'cromwell_user',
             'cromwell_password', 'MAX_CONTENT_LENGTH']
         for attr in requested_listener_attributes:
-            self.assertTrue(hasattr(config, attr), 'missing attribute %s' % attr)
+            self.assertTrue(hasattr(test_config, attr), 'missing attribute %s' % attr)
 
         # get an example wdl to test
-        wdl = config.wdls[0]
+        wdl = test_config.wdls[0]
         requested_wdl_attributes = [
             'subscription_id', 'wdl_link', 'workflow_name', 'wdl_default_inputs_link',
             'wdl_deps_link', 'options_link']
