--- listener-start.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-python green-box-api.py
