+++ requirements.txt
@@ -1,0 +1,8 @@
+wheel
+connexion
+google-auth
+google-cloud-storage
+flask
+requests
+git+git://github.com/broadinstitute/cromwell-tools.git
+git+git://github.com/HumanCellAtlas/pipeline-tools.git
